@@ -53,6 +53,49 @@ class PluginPackages:
 plugin_packages = PluginPackages()
 
 
+class AlphabeticalHelpFormatter(argparse.RawTextHelpFormatter):
+    """
+    Help message formatter with arguments sorted alphabetically.
+
+    This custom formatter extends RawTextHelpFormatter to sort command-line
+    arguments alphabetically when displaying help messages.
+    """
+
+    def add_arguments(self, actions):
+        """
+        Sort command-line arguments via alphabetical order when added to a parser.
+
+        This method takes a list of argument actions (positional or flag argument
+        strings) and sorts them alphabetically.
+
+        Parameters
+        ----------
+        actions : list
+            A collection of command-line argument definitions to be sorted.
+
+        Notes
+        -----
+        Sorting works as follows:
+
+        - For arguments with option flags (like -h or --help),
+          it uses the first option flag for ordering.
+        - For arguments without option flags, it uses the argument's name.
+        - For arguments with aliases flags (like config or conf),
+          it uses the non-alias action word for ordering.
+
+        Examples
+        --------
+        >>> argparse.ArgumentParser(formatter_class=SortingHelpFormatter)
+        >>> parser.add_argument('-z', '--zeta')
+        >>> parser.add_argument('-a', '--alpha')
+        >>> # Help text will show '-a, --alpha' before '-z, --zeta'
+        """
+        actions = sorted(
+            actions, key=lambda x: x.option_strings[0] if x.option_strings else x.dest
+        )
+        super().add_arguments(actions)
+
+
 class ParentParsers:
     """Object containing shared arguments for commands in a hierarchical order.
 
@@ -65,7 +108,7 @@ class ParentParsers:
     shared correctly.
     """
 
-    geoips_parser = argparse.ArgumentParser()
+    geoips_parser = argparse.ArgumentParser(formatter_class=AlphabeticalHelpFormatter)
     geoips_parser.add_argument(
         "-log",
         "--log-level",
@@ -76,9 +119,11 @@ class ParentParsers:
         help="Log level to output when using the CLI.",
     )
 
-    list_parser = argparse.ArgumentParser(add_help=False)
+    list_parser = argparse.ArgumentParser(
+        add_help=False, formatter_class=AlphabeticalHelpFormatter
+    )
     list_parser.add_argument(
-        "--package_name",
+        "--package-name",
         "-p",
         type=str,
         default="all",
@@ -103,6 +148,30 @@ class ParentParsers:
                 For more in formation on headers available, run
                 'geoips list <cmd> <positional_args> --columns help'.
                 """,
+    )
+
+    config_parser = argparse.ArgumentParser(add_help=False)
+    config_parser.add_argument(
+        "-n",
+        "--namespace",
+        default="geoips.plugin_packages",
+        type=str,
+        help=(
+            "The namespace of plugin packages to create plugin registries for. "
+            "If not specified, this defaults to 'geoips.plugin_packages'."
+        ),
+    )
+    config_parser.add_argument(
+        "-p",
+        "--packages",
+        default=None,
+        nargs="+",
+        type=str,
+        help=(
+            "The plugin packages to create or delete plugin registries for. Defaults to"
+            " None. If None, all plugin packages under 'namespace' will have their "
+            "plugin registries created or deleted, based on the command supplied."
+        ),
     )
 
 
@@ -162,6 +231,16 @@ class GeoipsCommand(abc.ABC):
             curr_parent = self.parent
             self.parent_parsers = []
             while curr_parent and hasattr(ParentParsers, f"{curr_parent.name}_parser"):
+                # Don't add parent parser arguments to 'geoips config install'. Config's
+                # shared arguments only apply to 'create-registries' and
+                # 'delete-registries' and the time being. I've looked for easier
+                # workarounds and can't find any suitable options other than hardcoding.
+
+                # If we end up needing to add more hardcoded special cases here, it
+                # might be worth looking into other methods of how to share arguments
+                # among specific commands, or subsets of commands. For now, this works
+                if self.name == "install" and self.parent.name == "config":
+                    break
                 self.parent_parsers.insert(
                     0,
                     getattr(ParentParsers, f"{curr_parent.name}_parser"),
@@ -195,7 +274,7 @@ class GeoipsCommand(abc.ABC):
                     parents=self.parent_parsers,
                     conflict_handler="resolve",
                     aliases=aliases,
-                    formatter_class=argparse.RawTextHelpFormatter,
+                    formatter_class=AlphabeticalHelpFormatter,
                 )
             except KeyError:
                 raise KeyError(
@@ -209,7 +288,7 @@ class GeoipsCommand(abc.ABC):
                 self.name,
                 conflict_handler="resolve",
                 parents=[ParentParsers.geoips_parser],
-                formatter_class=argparse.RawTextHelpFormatter,
+                formatter_class=AlphabeticalHelpFormatter,
             )
             self.LOG = self._get_cli_logger()
             self.combined_name = self.name
@@ -285,9 +364,11 @@ class GeoipsCommand(abc.ABC):
         """
         if len(self.command_classes):
             self.subparsers = self.parser.add_subparsers(
-                help=f"{self.name} instructions.",
+                help=f"{self.name} instructions."
             )
-            for subcmd_cls in self.command_classes:
+            # Sort subcommands alphabetically:
+            sorted_command_classes = sorted(self.command_classes, key=lambda x: x.name)
+            for subcmd_cls in sorted_command_classes:
                 subcmd_cls(LOG=self.LOG, parent=self, legacy=self.legacy)
 
     @property
